Move module runner functionality into server
